requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cssselect2==0.8.0
Flask==3.1.1
fonttools==4.58.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.10
langsmith==0.4.20
Markdown==3.8
MarkupSafe==3.0.2
mistune==3.1.3
numpy==2.3.1
orjson==3.11.3
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
beautifulsoup4==4.12.3
langchain-google-genai==2.1.10
langchain-openai==0.3.32
reportlab==4.4.2
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
slugify==0.0.1
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
zstandard==0.24.0
